+++ aqua/core/data_model/__init__.py
@@ -1,8 +1,4 @@
 from .coordidentifier import CoordIdentifier
 from .coordtransformer import CoordTransformer, counter_reverse_coordinate
 
-__all__ = [
-    "CoordIdentifier",
-    "CoordTransformer",
-    "counter_reverse_coordinate"
-]
+__all__ = ["CoordIdentifier", "CoordTransformer", "counter_reverse_coordinate"]